fix symlink error in getrootlessnetns
As described in #21194, even if i'ts not the root cause of the issue,
an error occurs when ~/.local/share/icontainers/storage is a link.

As it said in the #21194 comments, the rootless-ns code is rewritten in podman 5
so this patch will not be added in main project. I keep it to fix my
usage.

The function GetRoolessNetNs is called twice, in setUpNetwork and in
teardownNetworkBackend. The function use config.Engine.StaticDir to
calculate a hash to create the filename for setting NetNs.

In setUpNetwork, the value of config.Engine.StaticDir is the realPath
when the value is the link name in teardownNetworkBackend.

In teardownNetworkBackend, the netns file can't be found, the network is
not removed.

When a container is restarted or created with static ip, it fails with
existing ip error,

The patch replace the hash by a constant string.